Refactored auth blueprint
