+++ dribdat/app.py
@@ -51,6 +51,7 @@ def register_extensions(app):
 def register_blueprints(app):
     """Register Flask blueprints."""
     app.register_blueprint(public.views.blueprint)
+    app.register_blueprint(public.auth.blueprint)
     app.register_blueprint(public.api.blueprint)
     app.register_blueprint(user.views.blueprint)
     app.register_blueprint(admin.views.blueprint)
