+++ dribdat/public/views.py
@@ -2,32 +2,18 @@
 """Public section, including homepage and signup."""
 from flask import (Blueprint, request, render_template, flash, url_for,
                     redirect, session, current_app, jsonify)
-from flask_login import login_user, login_required, logout_user, current_user
+from flask_login import login_required, current_user
 
-from dribdat.extensions import login_manager, login_oauth
 from dribdat.user.models import User, Event, Project
-from dribdat.public.forms import LoginForm, UserForm, ProjectForm
-from dribdat.user.forms import RegisterForm
-from dribdat.utils import flash_errors
+from dribdat.public.forms import ProjectForm
 from dribdat.database import db
 from dribdat.aggregation import GetProjectData, ProjectActivity, IsProjectStarred, GetProjectTeam
 
-from datetime import datetime
-
 blueprint = Blueprint('public', __name__, static_folder="../static")
-
-def get_current_event():
-    event = Event.query.filter_by(is_current=True).first()
-    return event
-
-@login_manager.user_loader
-def load_user(user_id):
-    """Load user by ID."""
-    return User.get_by_id(int(user_id))
 
 @blueprint.route("/")
 def home():
-    event = get_current_event()
+    event = Event.current()
     if event is not None:
         events = Event.query.filter(Event.id != event.id)
     else:
@@ -36,128 +22,7 @@ def home():
 
 @blueprint.route("/about/")
 def about():
-    return render_template("public/about.html", current_event=get_current_event())
-
-@blueprint.route("/login/", methods=["GET", "POST"])
-def login():
-    form = LoginForm(request.form)
-    # Handle logging in
-    if request.method == 'POST':
-        if form.validate_on_submit():
-            login_user(form.user, remember=True)
-            flash("You are logged in.", 'success')
-            redirect_url = request.args.get("next") or url_for("public.home")
-            return redirect(redirect_url)
-        else:
-            flash_errors(form)
-    return render_template("public/login.html", current_event=get_current_event(), form=form)
-
-slack_oauth = login_oauth.remote_app("slack",
-    base_url = "https://slack.com/api/",
-    request_token_url = "",
-    access_token_url = "https://slack.com/api/oauth.access",
-    authorize_url = "https://slack.com/oauth/authorize",
-    consumer_key = "",
-    consumer_secret = "",
-    request_token_params = {"scope": "identity.email"},
-)
-
-@slack_oauth.tokengetter
-def slack_tokengetter():
-    return session.get("slack_token")
-
-@blueprint.route("/slack_oauth")
-def site_slack_oauth():
-    slack_oauth.consumer_key = current_app.config["DRIBDAT_SLACK_ID"]
-    slack_oauth.consumer_secret = current_app.config["DRIBDAT_SLACK_SECRET"]
-    return slack_oauth.authorize(
-        callback=url_for("public.slack_oauth_callback", _external=True)
-    )
-
-@blueprint.route("/slack_callback")
-@slack_oauth.authorized_handler
-def slack_oauth_callback(resp):
-    if resp is None or not resp["ok"]:
-        flash('Access denied to Slack', 'error')
-        return redirect(url_for("public.home"))
-    user = User.query.filter_by(sso_id=resp["user_id"]).first()
-    if not user:
-        if current_user and current_user.is_authenticated:
-            user = current_user
-            user.sso_id = resp["user_id"]
-        else:
-            user_data = slack_oauth.post("users.identity")
-            import string, random
-            rp = ''.join(random.SystemRandom().choice(string.ascii_uppercase + string.digits) for _ in range(20))
-            user = User.create(
-                username=user_data["user"]["name"].lower().replace(" ", "_"),
-                sso_id=resp["user_id"],
-                email=user_data["user"]["email"],
-                password=rp,
-                active=True)
-            user.socialize()
-            login_user(user, remember=True)
-            flash("Please complete your user account", 'info')
-            return redirect(url_for("public.user_profile"))
-    login_user(user, remember=True)
-    flash(u'Logged in via Slack')
-    return redirect(url_for("public.home"))
-
-@blueprint.route('/logout/')
-@login_required
-def logout():
-    """Logout."""
-    logout_user()
-    flash('You are logged out.', 'info')
-    return redirect(url_for('public.home'))
-
-
-@blueprint.route("/register/", methods=['GET', 'POST'])
-def register():
-    """Register new user."""
-    form = RegisterForm(request.form)
-    if request.args.get('name') and not form.username.data:
-        form.username.data = request.args.get('name')
-    if request.args.get('email') and not form.email.data:
-        form.email.data = request.args.get('email')
-    if request.args.get('web') and not form.webpage_url.data:
-        form.webpage_url.data = request.args.get('web')
-    if form.validate_on_submit():
-        new_user = User.create(
-                        username=form.username.data,
-                        email=form.email.data,
-                        webpage_url=form.webpage_url.data,
-                        password=form.password.data,
-                        active=True)
-        new_user.socialize()
-        if User.query.count() == 1:
-            new_user.is_admin = True
-            new_user.save()
-            flash("Administrative user created - have fun!", 'success')
-        else:
-            flash("Thank you for registering. You can now log in and submit projects.", 'success')
-        return redirect(url_for('public.login'))
-    else:
-        flash_errors(form)
-    return render_template('public/register.html', current_event=get_current_event(), form=form)
-
-@blueprint.route('/user/profile', methods=['GET', 'POST'])
-@login_required
-def user_profile():
-    user = current_user
-    form = UserForm(obj=user, next=request.args.get('next'))
-    if form.validate_on_submit():
-        originalhash = user.password
-        form.populate_obj(user)
-        if form.password.data:
-            user.set_password(form.password.data)
-        else:
-            user.password = originalhash
-        db.session.add(user)
-        db.session.commit()
-        user.socialize()
-        flash('Profile updated.', 'success')
-    return render_template('public/user.html', user=user, form=form)
+    return render_template("public/about.html", current_event=Event.current())
 
 @blueprint.route("/event/<int:event_id>")
 def event(event_id):
@@ -217,7 +82,7 @@ def project_new():
     project = Project()
     project.user_id = current_user.id
     form = ProjectForm(obj=project, next=request.args.get('next'))
-    event = get_current_event()
+    event = Event.current()
     form.category_id.choices = [(c.id, c.name) for c in project.categories_for_event(event.id)]
     form.category_id.choices.insert(0, (-1, ''))
     if form.validate_on_submit():
