+++ dribdat/templates/401.html
@@ -8,7 +8,7 @@
     <div class="text-center">
         <h1>401</h1>
         <p>You are not authorized to see this page. Please <a href="{{ url_for('public.home')}}">log in</a> or
-            <a href="{{ url_for('public.register') }}">create a new account</a>.
+            <a href="{{ url_for('auth.register') }}">create a new account</a>.
         </p>
     </div>
 </div>
