+++ dribdat/templates/nav.html
@@ -59,16 +59,16 @@
         </p></li>
       {% endif %}
         <li><p class="navbar-text">
-          [ <a class="navbar-link" href="{{ url_for('public.user_profile') }}">{{ current_user.username }}</a> ]
+          [ <a class="navbar-link" href="{{ url_for('auth.user_profile') }}">{{ current_user.username }}</a> ]
         </p></li>
         <li><p class="navbar-text">
-          <a class="navbar-link" href="{{ url_for('public.logout') }}"><i class="fa fa-sign-out"></i> Logout</a>
+          <a class="navbar-link" href="{{ url_for('auth.logout') }}"><i class="fa fa-sign-out"></i> Logout</a>
         </p></li>
     </ul>
     {% else %}
     <ul class="nav navbar-nav navbar-right">
-      <li><a href="{{ url_for('public.login') }}">Log in</a></li>
-      <li><a href="{{ url_for('public.register') }}">Create account</a></li>
+      <li><a href="{{ url_for('auth.login') }}">Log in</a></li>
+      <li><a href="{{ url_for('auth.register') }}">Create account</a></li>
     </ul>
     {% endif %}
   </div><!-- /.navbar-collapse -->
