+++ dribdat/templates/public/event.html
@@ -37,7 +37,7 @@
                 <br>project</span>
             </a>
         {% else %}
-          <p><a href="{{ url_for('public.register') }}">Register</a>
+          <p><a href="{{ url_for('auth.register') }}">Register</a>
           and log in to start a project.</p>
         {% endif %}
       </div>
