+++ dribdat/templates/public/login.html
@@ -16,6 +16,6 @@
   </form>
   <p><i>Please note, this is <b>not</b> the same login as for the Community board.</i></p>
   <p>Not yet registered?
-    <a href="{{ url_for('public.register') }}">Create an account</a>.</p>
+    <a href="{{ url_for('auth.register') }}">Create an account</a>.</p>
 </div>
 {% endblock %}
