+++ dribdat/templates/public/project.html
@@ -26,7 +26,7 @@
         <i class="fa fa-star-o"></i>JOIN
       </a>
       {% else %}
-      <a href="{{ url_for('public.login') }}" class="project-star" title="Login to join this project">
+      <a href="{{ url_for('auth.login') }}" class="project-star" title="Login to join this project">
         <i class="fa fa-star-o"></i>JOIN
       </a>
       {% endif %}
