+++ dribdat/templates/public/user.html
@@ -7,6 +7,6 @@
 {% block content %}
 <div class="container">
     <h2>My account</h2>
-    {{ render_form(url_for('public.user_profile'), form) }}
+    {{ render_form(url_for('auth.user_profile'), form) }}
 </div>
 {% endblock %}
